[base] [LINUXEP-7058] Change scheduled update behaviour so we check that the current time is within 2mins of scheduled time. If we update on a schedule, wait 5 minutes so we do not accidentally update twice. Added tests to test this functionality. Changed saltLength to a size_t to suppress warning that a signed integer was being compared to an unsigned.

+++ base/modules/Common/ObfuscationImpl/Cipher.cpp
@@ -98,7 +98,7 @@ namespace Common
         Cipher::Decrypt(const ObfuscationImpl::SecureDynamicBuffer& cipherKey, ObfuscationImpl::SecureDynamicBuffer& encrypted)
         {
 
-            int saltLength = encrypted[0];
+            size_t saltLength = encrypted[0];
 
             if(encrypted.size() < saltLength + 1)
             {
